Fix some sign comparison warnings.
R=tsepez@chromium.org

Review URL: https://codereview.chromium.org/1279493003 .

(cherry picked from commit b4c2167fff88665089eeab47afa45a9e890060a3)

Review URL: https://codereview.chromium.org/1287703002 .

+++ fpdfsdk/src/javascript/Document.cpp
@@ -1208,7 +1208,7 @@ CFX_WideString Document::CutString(CFX_WideString cbFrom) {
   wchar_t* pResult = (wchar_t*)malloc(iSize.ValueOrDie());
   wchar_t* pFrom = (wchar_t*)cbFrom.GetBuffer(iLength);
 
-  for (int i = 0; i < iLength; i++) {
+  for (size_t i = 0; i < iLength; i++) {
     if (pFrom[i] == L'\\' || pFrom[i] == L'/') {
       pResult[i] = L'\0';
       break;
